[refactor #15] AI 코드리뷰에 맞춰 디렉토리 구조 수정
