+++ src/main/java/com/example/login/common/exception/ApplicationException.java
@@ -1,6 +1,5 @@
 package com.example.login.common.exception;
 
-import com.example.login.common.errors.ErrorCase;
 import lombok.Getter;
 
 @Getter
