+++ src/test/java/com/example/login/application/service/UserServiceTest.java
@@ -2,7 +2,7 @@ package com.example.login.application.service;
 
 import com.example.login.application.dto.TokenDto;
 import com.example.login.application.dto.UserResponse;
-import com.example.login.common.errors.UserErrorCase;
+import com.example.login.common.exception.UserErrorCase;
 import com.example.login.common.exception.ApplicationException;
 import com.example.login.domain.user.entity.User;
 import com.example.login.domain.user.enums.Authority;
