+++ src/test/java/com/example/login/infrastructure/security/JwtTokenizerTest.java
@@ -1,7 +1,7 @@
 package com.example.login.infrastructure.security;
 
 import com.example.login.common.exception.ApplicationException;
-import com.example.login.common.errors.AuthErrorCase;
+import com.example.login.common.exception.AuthErrorCase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
